// migrations/snapshots/a40ed268abcc5dcf828ad3a7b53dda7bbdb8ca60508ed04ce9b280f161406dca/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'] | null;
      readonly thumbnail: CodecTypes['pg/text@1']['output'] | null;
      readonly isFeatured: CodecTypes['pg/bool@1']['output'];
      readonly status: CodecTypes['pg/bool@1']['output'];
      readonly views: CodecTypes['pg/int4@1']['output'];
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly tags: ReadonlyArray<
        | 'TECHNOLOGY'
        | 'LIFESTYLE'
        | 'TRAVEL'
        | 'FOOD'
        | 'FASHION'
        | 'BUSINESS'
        | 'ENTERTAINMENT'
        | 'SPORTS'
        | 'HEALTH'
        | 'POLITICS'
      > | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Profile: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly avatarUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly bio: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: Varchar<255>;
      readonly email: Varchar<255>;
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly activeStatus: 'ACTIVE' | 'BLOCKED';
      readonly role: 'USER' | 'AUTHOR' | 'ADMIN';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'] | null;
      readonly thumbnail: CodecTypes['pg/text@1']['input'] | null;
      readonly isFeatured: CodecTypes['pg/bool@1']['input'];
      readonly status: CodecTypes['pg/bool@1']['input'];
      readonly views: CodecTypes['pg/int4@1']['input'];
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly tags: ReadonlyArray<
        | 'TECHNOLOGY'
        | 'LIFESTYLE'
        | 'TRAVEL'
        | 'FOOD'
        | 'FASHION'
        | 'BUSINESS'
        | 'ENTERTAINMENT'
        | 'SPORTS'
        | 'HEALTH'
        | 'POLITICS'
      > | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Profile: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly avatarUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly bio: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
      readonly email: CodecTypes['sql/varchar@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly activeStatus: 'ACTIVE' | 'BLOCKED';
      readonly role: 'USER' | 'AUTHOR' | 'ADMIN';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly comment: {
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly isFeatured: CodecTypes['pg/bool@1']['output'];
      readonly status: CodecTypes['pg/bool@1']['output'];
      readonly tags: ReadonlyArray<
        | 'TECHNOLOGY'
        | 'LIFESTYLE'
        | 'TRAVEL'
        | 'FOOD'
        | 'FASHION'
        | 'BUSINESS'
        | 'ENTERTAINMENT'
        | 'SPORTS'
        | 'HEALTH'
        | 'POLITICS'
      > | null;
      readonly thumbnail: CodecTypes['pg/text@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly views: CodecTypes['pg/int4@1']['output'];
    };
    readonly profiles: {
      readonly avatarUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly bio: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly users: {
      readonly activeStatus: 'ACTIVE' | 'BLOCKED';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: Varchar<255>;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: Varchar<255>;
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly role: 'USER' | 'AUTHOR' | 'ADMIN';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly comment: {
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly isFeatured: CodecTypes['pg/bool@1']['input'];
      readonly status: CodecTypes['pg/bool@1']['input'];
      readonly tags: ReadonlyArray<
        | 'TECHNOLOGY'
        | 'LIFESTYLE'
        | 'TRAVEL'
        | 'FOOD'
        | 'FASHION'
        | 'BUSINESS'
        | 'ENTERTAINMENT'
        | 'SPORTS'
        | 'HEALTH'
        | 'POLITICS'
      > | null;
      readonly thumbnail: CodecTypes['pg/text@1']['input'] | null;
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly views: CodecTypes['pg/int4@1']['input'];
    };
    readonly profiles: {
      readonly avatarUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly bio: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly users: {
      readonly activeStatus: 'ACTIVE' | 'BLOCKED';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['sql/varchar@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly role: 'USER' | 'AUTHOR' | 'ADMIN';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;